add login/signup capabilities
