+++ controllers/api/index.js
@@ -3,7 +3,7 @@ const petImgRoutes = require('./petImgRoutes');
 const userRoutes = require('./userRoutes');
 
 
-router.use('/petimg', petImgRoutes)
-router.use('/users', userRoutes)
+router.use('/petimg', petImgRoutes);
+router.use('/users', userRoutes);
 
 module.exports = router;
