+++ models/User.js
@@ -1,6 +1,6 @@
 const { Model, DataTypes } = require('sequelize');
 const bcrypt = require('bcrypt');
-const sequelize = require('../controllers/connection');
+const sequelize = require('../config/connection');
 
 // create our User model
 class User extends Model {
