+++ public/assets/css/loginsignup.css
@@ -1,0 +1,3 @@
+.form {
+    align-items: center;
+}
