+++ public/assets/css/welcome.css
@@ -1,0 +1,31 @@
+.welcome {
+    align-items: center;
+    position: fixed;
+    width: 100%;
+    height: 100%;
+    left: 0;
+    display: flex;
+    justify-content: center;
+    flex-wrap: wrap;
+    background-image: url("https://thumbs.dreamstime.com/b/happy-dog-banner-copy-space-web-banner-happy-cute-jack-russell-terrier-puppy-pet-dog-125508754.jpg");
+    background-size: cover;
+    background-position: center;
+    background-attachment: fixed;
+    background-position: center;
+    z-index: 9999;
+}
+
+.redirect-text {
+    text-align: center;
+    color:rgb(198, 214, 198);
+    background-color: rgb(87, 102, 87);
+    border-radius: 15px;
+    padding: 20px;
+    border-style: solid;
+    border-color: rgb(41, 49, 41);
+}
+
+.redirect-text h1 {
+    padding: 35px 10px;
+    font-size: 40px;
+}
