+++ public/javascript/signup.js
@@ -1,0 +1,33 @@
+const signupForm = document.querySelector('#signup-form') || "";
+
+async function signupSubmit(event) {
+
+  event.preventDefault();  
+  console.log("hi");
+  const username = document.querySelector('#username-signup').value.trim();
+  const email = document.querySelector('#email-signup').value.trim();
+  const password = document.querySelector('#password-signup').value.trim();
+
+  if (username && email && password) {
+    const response = await fetch('/api/users', {
+      method: 'post',
+      body: JSON.stringify({
+        username,
+        email,
+        password
+      }),
+      headers: { 'Content-Type': 'application/json' }
+    });
+
+  if (response.ok) {
+      document.location.replace('/login');
+    } else {
+      alert(response.statusText);
+    }
+  }
+};
+
+signupForm && signupForm.addEventListener('submit', signupSubmit);
+
+
+console.log(signupForm);
