+++ views/comments.handlebars
@@ -5,3 +5,5 @@
     <p>{{this.comment_text}}</p>
 </div>
 {{/each}}
+
+{{!-- Make an option to add a comment --}}
