+++ views/layouts/main.handlebars
@@ -5,28 +5,40 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Petsagram</title>
-    <link rel="stylesheet" href="/stylesheets/style.css">
+    <link rel="stylesheet" href="/assets/css/main.css">
 </head>
 
 <body>
     <div class="wrapper">
-        <header>
-            <h1>
+        <header class="main-header">
+            <h1 class="welcome-header">
                 {{#if loggedIn}}
                 Welcome {{this.user.username}}!
                 {{else}}
                 Welcome to Petsagram!
                 {{/if}}
             </h1>
-            <nav>
-                {{#if loggedIn}}
-                <a href="/profile"></a>
-                <a href="/feed">feed</a>
-                <a href="/logout">Logout</a>
-                {{else}}
-                <a href="/login">Login</a>
-                <a href="/signup">Sign Up</a>
-                {{/if}}
+            <nav class="nav-header">
+                <ul>
+                    {{#if loggedIn}}
+                    <li>
+                        <a href="/profile"></a>
+                    </li>
+                    <li>
+                        <a href="/feed">feed</a>
+                    </li>
+                    <li>
+                        <a href="/logout">Logout</a>
+                    </li>
+                    {{else}}
+                    <li>
+                        <a href="/login">Login</a>
+                    </li>
+                    <li>
+                        <a href="/signup">Sign Up</a>
+                    </li>
+                    {{/if}}
+                </ul>
             </nav>
         </header>
         <main>
@@ -36,6 +48,8 @@
             {{!-- list our names or GitHubs here maybe --}}
         </footer>
     </div>
+    <script src="/javascript/login.js"></script>
+    <script src="/javascript/signup.js"></script>
 </body>
 
 </html>
