+++ views/signup.handlebars
@@ -1,5 +1,8 @@
 {{!-- Lets a user sign up, encrypts and saves the password. --}}
-<form class="signup-form">
+
+<link rel="stylesheet" href="/assets/css/loginsignup.css">
+
+<form class="form" id="signup-form">
   <div>
     <label for="username-signup">Username:</label>
     <input type="text" id="username-signup" />
@@ -13,6 +16,6 @@
     <input type="password" id="password-signup" />
   </div>
   <div>
-    <button type="submit">Sign Up</button>
+    <input type="submit"></input>
   </div>
 </form>
