+++ views/welcome.handlebars
@@ -1,5 +1,8 @@
 {{!-- Page that our site originally directs to. Should direct here every time that somebody tries to visit any other handlebars without trying to log in --}}
-
+<link rel="stylesheet" href="/assets/css/welcome.css">
 <div class="welcome">
-    <h1>Thank you for visiting Petsagram! Please log in or sign up so that you can explore and share!</h1>
+    <div class="redirect-text">
+    <h1>Thank you for visiting Petsagram!</h1> 
+    <h1>Please log in or sign up so that you can explore and share!</h1>
+    </div>
 </div>
